+++ tests/integration/test_kernel_maker.py
@@ -1,6 +1,7 @@
 """Tests for kernel_maker CLI module"""
 # Standard
 import argparse
+from datetime import datetime
 from unittest import mock
 # Installed
 import pytest
@@ -10,8 +11,10 @@ from libera_utils import kernel_maker
 from libera_utils.io.manifest import Manifest, ManifestType
 
 
+@mock.patch.object(kernel_maker, 'datetime', mock.Mock(wraps=datetime))
 def test_make_jpss_spk(test_data_path, short_tmp_path):
     """Test creating a SPK from packets"""
+    kernel_maker.datetime.utcnow.return_value = datetime(2025, 2, 25, 15, 45, 13)
     with mock.patch('libera_utils.spice_utils.KernelFileCache.cache_dir',
                     new_callable=mock.PropertyMock, return_value=short_tmp_path):
         packet_data_path = test_data_path / 'J01_G011_LZ_2021-04-09T00-00-00Z_V01.DAT1'
@@ -22,15 +25,17 @@ def test_make_jpss_spk(test_data_path, short_tmp_path):
             verbose=False
         )
         kernel_maker.make_jpss_spk(mock_parsed_args)
-        assert (short_tmp_path / 'libera_jpss_20210408t235850_20210409t015849.bsp').exists()
+        assert (short_tmp_path / 'libera_jpss_20210408t235850_20210409t015849_r25056154513.bsp').exists()
 
 
 @pytest.mark.parametrize(
     "wrapper",
     [AnyPath, S3Path, str]
 )
+@mock.patch.object(kernel_maker, 'datetime', mock.Mock(wraps=datetime))
 def test_make_jpss_spk_aws(test_data_path, short_tmp_path, create_mock_bucket, write_file_to_s3, wrapper):
     """Test creating a SPK from packets stored in AWS S3"""
+    kernel_maker.datetime.utcnow.return_value = datetime(2025, 2, 25, 15, 45, 13)
     bucket = create_mock_bucket()
     bucket = bucket.name
     key = 'some_path'
@@ -49,11 +54,13 @@ def test_make_jpss_spk_aws(test_data_path, short_tmp_path, create_mock_bucket, w
     kernel_maker.make_jpss_spk(mock_parsed_args)
 
     s3_output_path = S3Path(s3_output_directory)
-    assert (s3_output_path / 'libera_jpss_20210408t235850_20210409t015849.bsp').exists()
+    assert (s3_output_path / 'libera_jpss_20210408t235850_20210409t015849_r25056154513.bsp').exists()
 
 
+@mock.patch.object(kernel_maker, 'datetime', mock.Mock(wraps=datetime))
 def test_make_jpss_ck(test_data_path, short_tmp_path):
     """Test creating a CK from packets"""
+    kernel_maker.datetime.utcnow.return_value = datetime(2025, 2, 25, 15, 45, 13)
     with mock.patch('libera_utils.spice_utils.KernelFileCache.cache_dir',
                     new_callable=mock.PropertyMock, return_value=short_tmp_path):
         packet_data_path = test_data_path / 'J01_G011_LZ_2021-04-09T00-00-00Z_V01.DAT1'
@@ -64,15 +71,18 @@ def test_make_jpss_ck(test_data_path, short_tmp_path):
             verbose=False
         )
         kernel_maker.make_jpss_ck(mock_parsed_args)
-        assert (short_tmp_path / 'libera_jpss_20210408t235850_20210409t015849.bc').exists()
+        assert (short_tmp_path / 'libera_jpss_20210408t235850_20210409t015849_r25056154513.bc').exists()
 
 
 @pytest.mark.parametrize(
     "wrapper",
     [AnyPath, S3Path, str]
 )
+@mock.patch.object(kernel_maker, 'datetime', mock.Mock(wraps=datetime))
 def test_make_jpss_ck_aws(test_data_path, short_tmp_path, create_mock_bucket, write_file_to_s3, wrapper):
     """Test creating a CK from packets"""
+    kernel_maker.datetime.utcnow.return_value = datetime(2025, 2, 25, 15, 45, 13)
+
     bucket = create_mock_bucket()
     bucket = bucket.name
     key = 'some_path'
@@ -91,12 +101,14 @@ def test_make_jpss_ck_aws(test_data_path, short_tmp_path, create_mock_bucket, wr
     kernel_maker.make_jpss_ck(mock_parsed_args)
 
     s3_output_path = S3Path(s3_output_directory)
-    assert (s3_output_path / 'libera_jpss_20210408t235850_20210409t015849.bc').exists()
+    assert (s3_output_path / 'libera_jpss_20210408t235850_20210409t015849_r25056154513.bc').exists()
 
 
 @pytest.mark.xfail
+@mock.patch.object(kernel_maker, 'datetime', mock.Mock(wraps=datetime))
 def test_make_azel_ck(test_data_path, short_tmp_path):
     """Test creating a CK from packets"""
+    kernel_maker.datetime.utcnow.return_value = datetime(2025, 2, 25, 15, 45, 13)
     with mock.patch('libera_utils.spice_utils.KernelFileCache.cache_dir',
                     new_callable=mock.PropertyMock, return_value=short_tmp_path):
         packet_data_path = test_data_path / 'add-a-test-data-file.pkts'
@@ -107,7 +119,7 @@ def test_make_azel_ck(test_data_path, short_tmp_path):
             verbose=False
         )
         kernel_maker.make_azel_ck(mock_parsed_args)
-        assert (short_tmp_path / 'libera_azel_20210408t235850_20210409t015849.bc').exists()
+        assert (short_tmp_path / 'libera_azel_20210408t235850_20210409t015849_r25056154513.bc').exists()
 
 
 @pytest.mark.xfail
@@ -115,8 +127,11 @@ def test_make_azel_ck(test_data_path, short_tmp_path):
     "wrapper",
     [AnyPath, S3Path, str]
 )
+@mock.patch.object(kernel_maker, 'datetime', mock.Mock(wraps=datetime))
 def test_make_azel_ck_aws(test_data_path, short_tmp_path, create_mock_bucket, write_file_to_s3, wrapper):
     """Test creating a CK from packets"""
+    kernel_maker.datetime.utcnow.return_value = datetime(2025, 2, 25, 15, 45, 13)
+
     bucket = create_mock_bucket()
     bucket = bucket.name
     key = 'some_path'
@@ -135,12 +150,15 @@ def test_make_azel_ck_aws(test_data_path, short_tmp_path, create_mock_bucket, wr
     kernel_maker.make_azel_ck(mock_parsed_args)
 
     s3_output_path = S3Path(s3_output_directory)
-    assert (s3_output_path / 'libera_azel_20210408t235850_20210409t015849.bc').exists()
+    assert (s3_output_path / 'libera_azel_20210408t235850_20210409t015849_r25056154513.bc').exists()
 
 
+@mock.patch.object(kernel_maker, 'datetime', mock.Mock(wraps=datetime))
 def test_make_jpss_kernels_from_manifest_no_time_range(test_data_path, short_tmp_path):
     # Test that the kernels are generated when no desired range
     # is given.
+    kernel_maker.datetime.utcnow.return_value = datetime(2025, 2, 25, 15, 45, 13)
+
     m = Manifest(
         manifest_type=ManifestType.INPUT
     )
@@ -154,13 +172,16 @@ def test_make_jpss_kernels_from_manifest_no_time_range(test_data_path, short_tmp
     m.write(short_tmp_path, "libera_input_manifest_20220101t112233.json")
     kernel_maker.make_jpss_kernels_from_manifest(manifest_path, short_tmp_path)
 
-    assert (short_tmp_path / 'libera_jpss_20210408t235850_20210409t055849.bsp').exists()
-    assert (short_tmp_path / 'libera_jpss_20210408t235850_20210409t055849.bc').exists()
+    assert (short_tmp_path / 'libera_jpss_20210408t235850_20210409t055849_r25056154513.bsp').exists()
+    assert (short_tmp_path / 'libera_jpss_20210408t235850_20210409t055849_r25056154513.bc').exists()
 
 
+@mock.patch.object(kernel_maker, 'datetime', mock.Mock(wraps=datetime))
 def test_make_jpss_kernels_from_manifest_local_one_file(test_data_path, short_tmp_path):
     # Test that the kernels are generated when the desired range
     # falls within only one local file as in the example manifest file
+    kernel_maker.datetime.utcnow.return_value = datetime(2025, 2, 25, 15, 45, 13)
+
     m = Manifest(
         manifest_type=ManifestType.INPUT,
         configuration={
@@ -178,14 +199,17 @@ def test_make_jpss_kernels_from_manifest_local_one_file(test_data_path, short_tm
     m.write(short_tmp_path, "libera_input_manifest_20220101t112233.json")
     kernel_maker.make_jpss_kernels_from_manifest(manifest_path, short_tmp_path)
 
-    assert (short_tmp_path / 'libera_jpss_20210408t235850_20210409t015849.bsp').exists()
-    assert (short_tmp_path / 'libera_jpss_20210408t235850_20210409t015849.bc').exists()
+    assert (short_tmp_path / 'libera_jpss_20210408t235850_20210409t015849_r25056154513.bsp').exists()
+    assert (short_tmp_path / 'libera_jpss_20210408t235850_20210409t015849_r25056154513.bc').exists()
 
 
+@mock.patch.object(kernel_maker, 'datetime', mock.Mock(wraps=datetime))
 def test_make_jpss_kernels_from_manifest_local_two_files(test_data_path, short_tmp_path):
     # Test that the kernels are generated when the desired range
     # falls within two local files. This includes changing the time range
     # in the example manifest file and the expected output kernel names
+    kernel_maker.datetime.utcnow.return_value = datetime(2025, 2, 25, 15, 45, 13)
+
     m = Manifest(
         manifest_type=ManifestType.INPUT,
         configuration={
@@ -204,14 +228,17 @@ def test_make_jpss_kernels_from_manifest_local_two_files(test_data_path, short_t
 
     kernel_maker.make_jpss_kernels_from_manifest(manifest_path, short_tmp_path)
 
-    assert (short_tmp_path / 'libera_jpss_20210408t235850_20210409t035849.bsp').exists()
-    assert (short_tmp_path / 'libera_jpss_20210408t235850_20210409t035849.bc').exists()
+    assert (short_tmp_path / 'libera_jpss_20210408t235850_20210409t035849_r25056154513.bsp').exists()
+    assert (short_tmp_path / 'libera_jpss_20210408t235850_20210409t035849_r25056154513.bc').exists()
 
 
+@mock.patch.object(kernel_maker, 'datetime', mock.Mock(wraps=datetime))
 def test_make_jpss_kernels_from_manifest_local_three_files(test_data_path, short_tmp_path):
     # Test that the kernels are generated when the desired range
     # falls within three local files. This includes changing the time range
     # in the example manifest file and the expected output kernel names
+    kernel_maker.datetime.utcnow.return_value = datetime(2025, 2, 25, 15, 45, 13)
+
     m = Manifest(
         manifest_type=ManifestType.INPUT,
         configuration={
@@ -229,16 +256,19 @@ def test_make_jpss_kernels_from_manifest_local_three_files(test_data_path, short
     manifest_path = short_tmp_path / "libera_input_manifest_20220101t112233.json"
     kernel_maker.make_jpss_kernels_from_manifest(manifest_path, short_tmp_path)
 
-    assert (short_tmp_path / 'libera_jpss_20210408t235850_20210409t055849.bsp').exists()
-    assert (short_tmp_path / 'libera_jpss_20210408t235850_20210409t055849.bc').exists()
+    assert (short_tmp_path / 'libera_jpss_20210408t235850_20210409t055849_r25056154513.bsp').exists()
+    assert (short_tmp_path / 'libera_jpss_20210408t235850_20210409t055849_r25056154513.bc').exists()
 
 
+@mock.patch.object(kernel_maker, 'datetime', mock.Mock(wraps=datetime))
 def test_make_jpss_kernels_from_manifest_aws_three_files(test_data_path, short_tmp_path,
                                                          create_mock_bucket, write_file_to_s3):
     # Test that the kernels are generated when the desired range
     # falls within three remote files. This includes changing the time range
     # in the example manifest file and the expected output kernel names and
     # creating a mock bucket holding the data and the manifest file
+    kernel_maker.datetime.utcnow.return_value = datetime(2025, 2, 25, 15, 45, 13)
+
     bucket = create_mock_bucket()
     bucket_name = bucket.name
     key = 'some_path'
@@ -267,15 +297,18 @@ def test_make_jpss_kernels_from_manifest_aws_three_files(test_data_path, short_t
     s3_output_directory = S3Path(f"s3://{bucket_name}/{key}/kernel_output/")
 
     kernel_maker.make_jpss_kernels_from_manifest(manifest_path, s3_output_directory)
-    assert (s3_output_directory / 'libera_jpss_20210408t235850_20210409t055849.bsp').exists()
-    assert (s3_output_directory / 'libera_jpss_20210408t235850_20210409t055849.bc').exists()
+    assert (s3_output_directory / 'libera_jpss_20210408t235850_20210409t055849_r25056154513.bsp').exists()
+    assert (s3_output_directory / 'libera_jpss_20210408t235850_20210409t055849_r25056154513.bc').exists()
 
 
+@mock.patch.object(kernel_maker, 'datetime', mock.Mock(wraps=datetime))
 def test_make_jpss_kernels_from_manifest_aws_one_file(test_data_path, short_tmp_path,
                                                       create_mock_bucket, write_file_to_s3):
     # Test that the kernels are generated when the desired range
     # falls within one remote file. This includes creating a mock bucket
     # holding the data and the newly made manifest file also in the bucket
+    kernel_maker.datetime.utcnow.return_value = datetime(2025, 2, 25, 15, 45, 13)
+
     bucket = create_mock_bucket()
     bucket_name = bucket.name
     key = 'some_path'
@@ -294,5 +327,5 @@ def test_make_jpss_kernels_from_manifest_aws_one_file(test_data_path, short_tmp_
     s3_output_directory = S3Path(f"s3://{bucket_name}/{key}/kernel_output/")
 
     kernel_maker.make_jpss_kernels_from_manifest(manifest_path, s3_output_directory)
-    assert (s3_output_directory / 'libera_jpss_20210408t235850_20210409t015849.bsp').exists()
-    assert (s3_output_directory / 'libera_jpss_20210408t235850_20210409t015849.bc').exists()
+    assert (s3_output_directory / 'libera_jpss_20210408t235850_20210409t015849_r25056154513.bsp').exists()
+    assert (s3_output_directory / 'libera_jpss_20210408t235850_20210409t015849_r25056154513.bc').exists()